+++ src/engraving/rendering/editmode/editmode.cmake
@@ -3,6 +3,4 @@
 set(RENDERING_EDITMODE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/editmoderenderer.cpp
     ${CMAKE_CURRENT_LIST_DIR}/editmoderenderer.h
-    ${CMAKE_CURRENT_LIST_DIR}/editmodedraw.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/editmodedraw.h
 )
